Add classs into separate files
